Fixes exit(1) when accessing placeholder typedef with nonexistent type
Summary:
In the Thrift compiler, when you declare something (field, const, etc.) with a type that is not yet known, the parser creates a placeholder typedef that stores only the symbolic name of the type. Later in the process, when you call `get_type()` on the `t_typedef` AST node to retrieve the actual type, a scope lookup is carried out to find the actual type.

However, if a type of that name cannot be found, calling `get_type()` on a `t_typedef` currently does an `exit(1)` (!!).

To address this, we do a sweep of all generated placeholder typedefs after parsing each Thrift file to resolve the target types. If any of the generated placeholder typedef refers to non-existent type, we issue a compilation failure message. This ensures that `get_type()` would never be called with an nonexistent type. We can therefore replace the `exit(1)` check there with an `assert`.

Note: I intentionally chose a pretty verbose name `named_placeholder_typedefs` for tracking placeholder typdefs. The memory leak patching diff stack actually adds this field with the name `placeholder_typedefs`. I chose a different, non-obvious name here to ensure that I get a merge conflict when that stack lands.

Note 2: Since the check is done after regular parsing, it will always report the lint at the end of the file. We can change `failure` (and friends) to take an optional line number override, but that should probably be its own diff.

Reviewed By: yfeldblum

Differential Revision: D9473401

fbshipit-source-id: 5eabc83d3838619272ec24ed6cafc2fb51bc488b