Add an annotation to minimize padding of generated C++ structs
Summary:
Add `cpp.minimize_padding` Thrift struct annotation that enables reordering of member variables in generated C++ structs to minimize padding.

For example, the Thrift struct

```
struct foo {
  1: required byte small,
  2: required i64 big,
  3: required i16 medium,
  4: required i32 biggish,
  5: required byte tiny,
} (cpp.minimize_padding)
```
is converted into C++ as if member order was

```
  2: required i64 big,
  4: required i32 biggish,
  3: required i16 medium,
  1: required byte small,
  5: required byte tiny,
```

This is done by sorting members in the order of decreasing estimated alignment.

This should work with the `optionals` setting which doesn't change field alignments (at least in the `folly::Optional` implementation).

Reviewed By: yfeldblum

Differential Revision: D9353622

fbshipit-source-id: eecd40e5ab78e9372f201a8a8e42b7a7557c7793