Implements t_program_bundle to own all t_program-s
Summary:
Part of the grand scheme to patch memory leaks in the Thrift parser library.

When we parse a Thrift file, we create a root `t_program` (the file that we parsed), as well as potentially many other `t_program`s that correspond to included files. This diff adds a `t_program_bundle` construction to own all the `t_program`s.

Most of the changed files are callsite changes.

Reviewed By: yfeldblum

Differential Revision: D9467063

fbshipit-source-id: 59c93e02fb5622a189922d5bf13faa7479e45ea7