Fixes orphaned AST nodes created during the include pass
Summary:
Part of the grand scheme to patch memory leaks in the Thrift parser library.

Thrift parsing process consists of two passes: first to collect all includes, and then to actually construct the AST. Therefore, for objects created in the first pass, their ownership is not actually passed to the program AST, and they are therefore orphaned. Note that they can not be deleted on the spot 'cause some of them are referenced later in the parsing process.

Therefore this diff creates a mechanism to pass these objects to the `parsing_driver` for deletion when the `parsing_driver` is destructed.

Reviewed By: yfeldblum

Differential Revision: D9467701

fbshipit-source-id: 72dc21e92be7429c003fa3ac0c6e97981b3f3339