Eliminates placeholder typedef leaks
Summary:
Part of the grand scheme to patch memory leaks in the Thrift parser library.

There are two (maybe more) types of typedef-s: 1) explicit `typedef ...` in the Thrift program, and 2) a "placeholder" typedef created when the program references a type that has not yet been encountered. In the case of 1), the typedef is gleefully owned by its corresponding `t_program`. Currently, however, typedef-s created in case 2) are not owned by anyone.

This diff makes `t_program` own these placeholder typedef-s as well.

Reviewed By: yfeldblum

Differential Revision: D9467703

fbshipit-source-id: 9fbb759a0dc9719158e006e4782cf051ad054e2d