containers element/key aliasing
Summary:
Before this diff we sometimes inadvertenly created copies of
elements or keys in containers, when accessing them or doing comparisons.

These changes make sure that no copies take place in the cython wrappers
themselves.

This follows in line with the changes to Structs and Constants to insure
we use aliasing instead of copies.

Reviewed By: yfeldblum

Differential Revision: D9506660

fbshipit-source-id: a3a799135be7cdef6f5281ee02053869e3aa71d9