Check HTTP status code in HTTP2 transport
Summary:
The HTTP2 transport doesn't check the HTTP status code (like the HTTP1
transport does), and so failed requests result in obscure errors like
TProtocolException trying to decode an HTML response. This changes it to throw
an exception if the code is not 100 or 200 (same as HTTP1 logic).

I've also improved unknown channel errors to include the Proxygen error code,
and made the no content exception more descriptive.

Reviewed By: yfeldblum

Differential Revision: D8661729

fbshipit-source-id: 8e717370543857dc89a80eab94bd59f016dfac8d