Fixup rich comparisons, support cases where some structs are unorderable
Summary:
Modified the compiler to make avaiable a function from the cpp2 generator,
to determine if a type is orderable, so we could provide ordering comparison
operators if a Struct is orderable in cpp2.

Also fixed up the rich comparisons for thrift-py3 List's so its easier
to read and supports ordering like python lists.

Cleaned up the richcmp for set and dict to make use of #defines from cpython for
the comparison type instead of using magic numbers.

Reviewed By: yfeldblum

Differential Revision: D9327740

fbshipit-source-id: a6cac88769a0124bdc3bc882287dd17c008feace