Add the Serious exception to the thrift fixtures
Summary:
We are updating the hack codegen for exceptions with a `getMessage()` method in cases where a message annotation is present.

This diff adds a third exception to the exception fixtures module in the thrift codegen. This exception uses a `message` annotation that references a field that is not `message`. At least in Hack the `message` field of the struct overrides the `message` variable of the parent exception so observing the behavior of the codegen in both scenarios is important.

I don't understand what the `Banal` and `Fiery` names are supposed to mean so I struggled with a name for the third exception. In Romanian `Banal` means trivial, `Fiery` might be that it is triggered (fires) often. In those lines I picked Serious in the idea that this is the important exception on our service to pay attention to. I welcome any name suggestions.

Reviewed By: yfeldblum

Differential Revision: D9499519

fbshipit-source-id: d538a5f10a9d4d23854b4e1cb09791bafac13c4e