Add message annotation support for thrift codegen in hack
Summary:
We are updating the hack codegen for exceptions with a `getMessage()` method in cases where a message annotation is present.

The main decision for this change are what to do when the value of the `message` annotation is `message` or when there is no annotation.
The current behavior is to return the `message` field value from `getMessage()` when the field is present so overriding `getMessage()` to return an empty string in those scenarios is problematic at this time - we would want to add the `(message = "message")` annotation to all the thrift exceptions before that change.
When the annotation points to the `message` field it is less clear what to do. The `message` variable on Exception is protected in Hack and `getMessage()` will return the contents of the message. The main risk is Hack changing the visibility of the message variable to private and that is actually a reasonable thing to do. In the end I think that it is better to generate the `getMessage()` method even if the value of the annotation is `message`.
Another issue to address is what is the behavior when the annotation references a field that is not required. In hack the struct generates a nullable field. It makes most sense to me to return an empty string when the field is null.

Reviewed By: yfeldblum

Differential Revision: D9499518

fbshipit-source-id: 0845d197c8ba2d3d6d1d7302629e7fef46c4a07f