Skip generating getMessage when the annotation is message
Summary:
When we have an exception with the `message` field, the hack codegen will define the fields as a non nullable message as opposed to the standard nullable field. This makes the conversion to string using `?? ''` not required for the `message` fields. A secondary issue is that the base exception class in hack is whitelisted in Autobahn where the methods in classes that extend getMessage are not.

When we don't override the getMessage method for the annotations with the message field we preserve enough backwards behavior that things work in WWW.

Previous changes are in D9499518.

Reviewed By: spalamarchuk

Differential Revision: D9687075

fbshipit-source-id: 6a5e581171b7abdd2f6117c6feb459995d0e6d69