Serialization/deserialization for Thrift rsocket-protocol client and server
Summary:
This diff lays the basic foundation needed for the new Thrift rsocket
client and server. All the needed frame types are introduced, along with code
to serialize/deserialize the frame structures.

Note that in this first iteration, serialization performance is not a focus,
though I plan to improve serialization performance in the future. (For
starters, there is no point in allocating the small frame headers on the heap.)

Reviewed By: yfeldblum

Differential Revision: D9326507

fbshipit-source-id: 1e4fd6022fee2ea1dee08b811aa6c709ead78ccf