Persist a Cpp2ConnContext for the life of a Thrift rsocket server connection.
Summary:
The present Thrift rsocket server implementation creates a new
Cpp2ConnContext for each request, not once for a connection.
This change creates a Cpp2ConnContext on the first request and reuses
it as the value of Cpp2RequestContext::ctx_ in subsequent requests.
This change minimizes source changes, at the cost of being less elegant
than one which changes more APIs and affects more transports.

Reviewed By: andriigrynenko

Differential Revision: D9755365

fbshipit-source-id: b978f9350763e5397c8936b91b8970eb4ea1f6a0