Use defined namespace and not local namespace for enums in cocoa generator (#291)
Summary:
Referencing an enum from a different namespace in a thrift file does not use the referenced namespace as prefix in the cocoa generator. Currently it uses the namespace of the current thrift. This diff tries to fix this issue by prefix the the enum with the namespace where the enum is actually defined instead of using the local one.
Pull Request resolved: https://github.com/facebook/fbthrift/pull/291

Reviewed By: vitaut

Differential Revision: D9927755

Pulled By: yfeldblum

fbshipit-source-id: 2e558573e6491bf381d81f0011cefa56525ad209