improve struct._make_instance performance of struct.__init__ by ~50% {FIXED}
Summary:
Fixed bug in not being able to initialize cython arrays, so we set True/False
instead of just True.

A few things, we where calling everything 'objects' in make_instance so
we could accept 'NOTSET'.  This ment that none of the speed of Cython could be
used because it didn't know the real types.

So to encode a string instead of using PyUnicode_toUTF8String() cython was
having to call a bunch of python code, to call the 'encode' method on
what ever the field happened to be.

I removed the need to pass NOTSET into make_instance, by instead using
an array of flags to indicate the field was not set during a __call__

Also Cython doesn't assume C++11 so it can't use move(). It turns
out because of exception handling it copies the std::string an extra time
when assigning the value to the cpp2 struct.  So we use move() in thrift-py3
and the .cpp code is correct to move instead of copy.

Also replaced the built in cython bytes to std::string code with one that
is marked "inline" seems we see some good improvements there as well.

This will make `struct.__call__` faster as well but I was only profiling `__init__`

Reviewed By: yfeldblum

Differential Revision: D10138630

fbshipit-source-id: 69f9c481c3950a9360a537e1c6244b7148aeb672