Fix exception type support in cocoa generator (#290)
Summary:
Currently it seems it's not possible to create an exception type via fbthrift. For example if you declare the following exception:

```
exception RetriableRPCException {
    1: i32 errorCode,
    2: string message,
}
```

The cocoa generator creates the same output as a struct except it inherits from `NSException` instead of `TBaseStruct`. The problem is that it is adding getter and setter that are calling immutable methods as well as the `makeImmutable` method that all are only implemented in `TBaseStruct`. There are various approaches to address this, but the one taken here is to make a `TBaseException` type and clone the immutability interface from `TBaseStruct` into `TBaseException`.

Pull Request resolved: https://github.com/facebook/fbthrift/pull/290

Reviewed By: spalamarchuk

Differential Revision: D10231684

Pulled By: yfeldblum

fbshipit-source-id: 845f004bdb69c1b7f346c9d40e8ea15473a90feb