more name mangling for struct private cache names
Summary:
If a field is named hash, and it happens to be a container/struct/iobuf
we keep a local cache copy of the wrapping python object.

We were calling these local copies `__{{field:name}}`,  well we also had
a `__hash` to store the cache hash() value for the struct.

These two collide, so I add `field_` to all field cache attributes.

(Note: this ignores all push blocking failures!)

Reviewed By: yfeldblum

Differential Revision: D10282576

fbshipit-source-id: 9d7614e2a13826a9a0060db3ed5b83982b60e991