fix py3.flags to not try to pre-compute combination enum members
Summary:
Was not thinking when I tried to precompute combination flags in py3.flags
enums.  This is pretty rediculous as 32bits of flag combinations is quite large

The cpu cost to decompose a flag combination is always O(n) where n < 32.
So instead of pre-computing, just decompose the value in real time.

We still cache them so that we will never have an enum instance with
the same values but different `id()` thus failing the `is` check

Reviewed By: yfeldblum

Differential Revision: D10391339

fbshipit-source-id: d7539e4e5ebfa455645cb4d5e250e9770d417aaa