Hack, throw when serializing struct with missing required fields
Summary:
Currently when serializing a thrift struct, we skip any field with a value equal to `null`, even when the field is required.

This diff, introduce a change of the `$_TSPEC` datastructure to add information about the "requiredness" of the field. The piece of information is then later used during the serialization, if the field is `null` and `required` then we throw an `TProtocolException`.

Reviewed By: vitaut

Differential Revision: D10282337

fbshipit-source-id: 4fbeb14c24f39a2c123fbe9fac5ae2d3110f8603