Optimize TSimpleJSONProtocol.writeJSONString()
Summary:
writeJSONString() can be very expensive in json serialization
due to the many calls to writeJSONChar() and the isinstance
checks done there.

The checks are not needed, we always pass a char to that
function. Also when possible, we can iterate the string
instead of using range, that is slightly faster.

Reviewed By: yfeldblum

Differential Revision: D10406047

fbshipit-source-id: 0de209b5c54c50b6c51875f495be00a08a863d52