optimize TSimpleJSONProtocol.writeJSONString()
Summary:
Writing to the buffer one char at a time really hurts performance. Most strings
in thrift don't require escaping, so this optimizes for that.

In some workloads this optimization reaches 10x speed improvement.

This prepares a set of chars (both bytes and unicode variants) that need
escaping for fast lookups, peeks into the string to see if escaping is required,
and fast-circuits the writes of the chunks that require no escaping.

Reviewed By: yfeldblum

Differential Revision: D10454647

fbshipit-source-id: 8d83df834bf72bb1d504c35f558967786f248012