add a new thrift.util.inspect python module
Summary:
This module contains helper functions for introspecting thrift services from
python.

This makes it easier for users to implement their own custom versions of the
thrift pyremote scripts.  The pyremote script code currently relies on the
thrift compiler to emit a skeleton file containing the function signature
information.  Unfortunately this skeleton is not really set up to be used as a
library.

This new inspect module allows this information to be extracted automatically
from the thrift_spec structures already emitted in the service library module
generated by thrift.

Reviewed By: yfeldblum

Differential Revision: D10324858

fbshipit-source-id: 1047b28446970092b0dedba8ffee72dd2e8e9ae9