Set SR headers, keep connection alive and flush before closing
Summary:
Call `Flush()` before closing the connection, so to avoid unneeded EOFs.

When an application generates an undeclared exception (as defined in the thrift spec), allow it to be forwarded to the client without closing the connection. This also fixes the case the it's a "fire & forget" call, which must not generate exceptions, as per the thrift IDL spec.

Reviewed By: lclarkmichalek

Differential Revision: D10424318

fbshipit-source-id: ac3ec84ac11b8109ed894e2c4c2d00e34ca0a70f