cmake in a _build subdir to fix macos build
Summary:
This addresses a hilarious but hard to trace build problem on
macOS.  The issue is that the build manages to resolve `folly/String.h`
in place of `string.h` and breaks compilation of `<cstring>` and all
that include it.

The resolution is to ensure that we generate a subdir for the cmake
build.

Reviewed By: simpkins

Differential Revision: D10520960

fbshipit-source-id: 9bebeda69c6e28ebca146f9b96ee01bedd2d6ef1