wrap includes of Kerberos headers in `extern "C"`
Summary:
The Kerberos headers are C-only, so wrap them in `extern "C"` when including
them.

In particular, `krb5/krb5.h` includes `comm_err.h`, and depending on the
platform this header does not always correctly guard itself with `extern "C"`
when it is included from C++ code.  For instance this is the case on Ubuntu
18.04, which resulted in the thrift code incorrectly treating the functions
defined in `com_err.h` as having C++ linkage.

Reviewed By: yfeldblum

Differential Revision: D10498230

fbshipit-source-id: 353443e4172285d6510860a8a5a3afcebc190e00