Preserve old serialization format with the new folly::none
Summary:
Changing folly::none to a type that could not be implicitly converted
to bool subtly changed the serialization logic in
thrift::frozen. Unset Optional<bool> used to be serialized as false,
not none. After D10475780, it would be serialized as none, which
changed the serialization format.

Introduce a new none-like type that preserves that old behavior.

Reviewed By: yfeldblum

Differential Revision: D12873924

fbshipit-source-id: 043b1052f2404dc7a5638a4f774c11bc5982e4ec