Add `setStartedProcessing` hook for the ResponseChannelRequest
Summary:
Add a `setStartedProcessing` hook in `ResponseChannelRequest` that is called as soon as we start processing the request.

Currently that hook is not doing anything, but it will later be used to hook the admission control mechanism.

The destructor of `ResponseChannelRequest` will (later) also be used to indicate the end of processing (i.e. after we send the response to the client).

Reviewed By: yfeldblum

Differential Revision: D8953107

fbshipit-source-id: 4e899051f0b8127fdac1160277e143431a2473f1