improve error messages from the compiler fixture tests
Summary:
If differences are found in one of the output files, fail with a unified diff
of the generated file and its expected contents.  Also include the filename in
the error message.

Previously the full contents of the file were printed, which was pretty
awkward for one-line differences in files that were often over 1000 lines
long.  The test code did print the filename on a line by itself to stderr, but
this was typically hard to find since it was shown separately from the failure
message.

Reviewed By: yfeldblum

Differential Revision: D13038238

fbshipit-source-id: c14e896152e8281718e44922041861ee152ca967