Introduce AdmissionController/Strategy
Summary:
Introduce the Admission controller abstractions (currently compiled but not used).

The AdmissionController API is composed of 3 methods (`admit`, `dequeue` and `response`) which are called at different time in the processing of a request.
- `admit` is called after we received a frame, before deserializing the content of the request. It return a boolean which indicates if we are admitting the request (if false we respond with a loadshed exception)
- `dequeue` is called when we start processing the request.
- `returnedResponse` is called when we have finished responding to the request

The AdmissionStrategy role is to assign an AdmissionController to a request. In the simplest implementation, all requests use the same AdmissionController, but we can also have per clientId controller or even priority aware ones.

Reviewed By: spalamarchuk

Differential Revision: D8955476

fbshipit-source-id: d4c9b3cd5a95bbccbd778c84edaa6633f27f1335