Kill off some shared_ptr refcount modifications.
Summary:
The auto shared_ptr in a for loop strikes again. We iterate over the vector,
doing an atomic RMW on each element's refcount.

Since concurrent modification of the vector contents is a bug anyways, and
since we're not storing the shared_ptrs themselves anywhere, we can take them
by reference.

Reviewed By: jmswen

Differential Revision: D13309052

fbshipit-source-id: 6428aa69ceb6b81d6667c26e7b99b9ffc7db581f