Embed tls.ConnectionState in server contexts
Summary:
Update the go thrift library to embed TLS information in the ConnInfo
in server contexts. This allows handlers to capture client TLS certificates.

There's also a bug fix in the compiler in the code generation using contexts,
and an update to the doc string to expose the 'use_context' option.

Differential Revision: D13319151

fbshipit-source-id: e89f9b80a829498ad3d87f8a164aa868e9f919df