(thrift/rocket)(rfc) reducing memory footprint of Stream object
Summary:
When using Yarpl, chaining a new operator (e.g. map/observeOn) may consume 200 ~ 300 bytes of memory. (I did a test to write a trivial operator which did nothing and it takes 200 bytes). The idea of this diff is to avoid create a chain of operators and only use a thin layer of Yarpl (one flowable + subscriber + subscription). This could cut down the memory usage from 1.2~1.6 KB to 300~400 KB when using rocket client stream.

I think the trade off here is that it will lose the capability to dynamically chaining arbitrary operators as Yarpl provide. But seems that might be something don't need here, as Stream provides a pretty tight interface here. However things could change if extending Stream interface to be more flexible.

This is still a very preliminary diff. This implementation is not very much thread safe and I didn't do an equivalent of EagerSubscribeOnOperator in YarplStreamImpl. Would like to hear folks idea of whether this is a good direction to try out or out.

Reviewed By: phoad

Differential Revision: D12887515

fbshipit-source-id: 1e957987cd622d6a6442ac9c894b6c0f27efcec4