Rename param in code generated by the Go Thrift compiler
Summary:
We have a thrift file that uses `r` as a parameter name, causing a conflict with the code generated by the Go thrift compiler, which also uses `r` for a named return value.

My first plan was to eliminate named return values in the Go compiler, but this is a bit more involved. I would like to first unblock and then implement the real fix. This diff just renames the parameter.

Reviewed By: mkorenkov

Differential Revision: D13548232

fbshipit-source-id: 35600ee9df92150045cb38677c33f3425ff54718