Add a few tests exercising RocketClient shutdown logic
Summary: This diff does two things: 1) ensures that all streams in `RocketClient` unit tests are thread-safe and eagerly subscribed to by ensuring we have the `YarplStreamImpl` layer; and 2) adds a few more tests that stress client shutdown logic. These two items are related: one of the new tests will fail without `YarplStreamImpl` in the picture, since if a flowable is not eagerly subscribed to, the `onSubscribe()` callback may attempt to use the underlying `RocketClient` after it has been destroyed.

Reviewed By: stevegury

Differential Revision: D13522018

fbshipit-source-id: b5bb7823742cdbae5f222e5e4e2482b2ae05ace5