Enable copy_from [optional_]field_ref of another type
Summary:
This diff enables `copy_from` for `[optional_]field_ref` referring to a different
type, e.g.
```
a.foo_ref().copy_from(b.bar_ref());
```
where `a` is a Thrift object with field `foo` of type `T` and `b` is a Thrift
object with field `bar` of type `U` assignable to `T`.

Reviewed By: yfeldblum

Differential Revision: D13530841

fbshipit-source-id: 85e3ab50937ca6ec1b2999a991bc544c26507ff3