Improve Frames handling in RocketServerConnection
Summary:
The current frame handling logic on server is a bit complicated and does unnecessary stuff on fast path (when a whole request fits into a single frame), in particular it does move the frame around and goes via boost visitor.

This diff simplifies that a bit, in particular the `RocketServerFrameContext` is split into two classes  `RocketServerFrameContext` (now doesn't hold any temporary data) and `RocketServerPartialFrameContext` (holds temporary data needed for fragments handling). On fast path the logic is now linear with no visitors, slow path is also improved a bit.

Before this diff, I could observe ~6% of CPU spent between the call to `RocketServerConnection::handleFrame` and the call to `ThriftRocketServerHandler::onFullFrame`. After this diff, this drops to ~3%.

Reviewed By: jmswen

Differential Revision: D13837724

fbshipit-source-id: 711b6e1c532d95036f49690afde49893aeb3a325