Thrift: Don't force coroutine implementations
Summary:
Initially, coroutine-tagged thrift methods had to be implemented with C++
coroutines. This can be a hassle if you need to implement the same API with two
different interfaces, only one of which wants to be coroutine-ified (or only
one of which builds on a compiler that supports coroutines). This allows such
methods to be implemented either with coroutines, or with another mechanism.

Reviewed By: andriigrynenko

Differential Revision: D13830550

fbshipit-source-id: 2bc4ff84d6a2840d81acdb16a1695e808f308bd5