Fix concurrency bug in QIAdmissionController
Summary:
Right now we grab the current time outside of the lock.
This can lead to computing a negative time difference.

Example, two threads can (in order):
T2 grabs the time `t0` in `dequeue`
T1 grabs the time `t1` in `admit`  (`t0 < t1`)
T2 waits on mutex
T1 continues and update the lastTime to `t1`, unlock the mutex
T2 continues and compute `elapsed = now - lastTime`, i.e. `t0 - t1` which is negative

To solve this problem, I move back the call to `now()` after the mutex.

Note: `returnedResponse()` still grabs the current time before the mutex, since it doesn't really matter to have the exact current time to insert the value in the current bucket of the time series.

Reviewed By: yfeldblum

Differential Revision: D13848183

fbshipit-source-id: 5db57284f27ee49a679426c4ee7c4482bc45d8fa