Avoid double parsing of frame headers
Summary:
We currently first read the header to decide what is the streamid / frame type and then read it again when constructing the frame.

This diff removes the duplication and simply forwards available information along with the cursor to the Frame constructors.

I could observe this giving ~0.4% perf improvement in my loadtest.

Reviewed By: yfeldblum

Differential Revision: D13873350

fbshipit-source-id: 30b67668832b0cf1f43d21bd244dadd110db3a50