limit badness of pathological hash table case
Summary:
If a frozen hash table is used in a context where there is a large common
divisor between integer keys and the hash table bucket count, lookups can
become pathologically bad. The bucket count is a multiple of 64, so this
is particularly likely to happen for powers of two.  This diff limits the
pathology to be only somewhat bad by ensuring that the bucket count for
newly created frozen tables is not divisible by 128.  In this scenario
average probe lengths can still be 12.8 instead of the desired 1.6,
but without it they have been observed in production to be more than 100.

A similar issue can occur if the bucket count is a multiple of 5, due to
a multiplier applied to the hash value.  Excessive probing can similarly
be avoided by slightly increasing the bucket count.

Getting rid of the pathology more completely should certainly be done,
but the need to maintain forward and backward compatibility for frozen
structures makes it challenging.

This diff will increase the size of frozen hash tables by 14.4 bytes
on average. Small tables (less than 25 entries) will not have their
size affected.

Reviewed By: ddrcoder

Differential Revision: D13934860

fbshipit-source-id: 7256320b89f1da34e7e873ab7db12b38c96908cc