Fix JSON serialization of nested structs in Python
Summary:
Fix context handling in `TJSONProtocol`. Previously the current context wasn't updated on pop which, in particular, resulted in the following invalid JSON output:

```
{"1":{"rec":{}}"2":{"rec":{}}}
```
(note missing comma between `}` and `"2"`).

With this fix (controlled by `validJSON` parameter to `TJSONProtocolFactory` and disabled by default) the output is correct:

```
{"1":{"rec":{}},"2":{"rec":{}}}
```

In preparation for rolling out the fix, make `TJSONProtocol` in Python accept valid JSON.

Reviewed By: stevegury

Differential Revision: D13883630

fbshipit-source-id: 3adf5190599686979e6121428d4f84839ccea0ca