Kill responseSizeChecked_ in ThriftRequestCore
Summary:
From what it appears, the purpose of this flag is to avoid recursion `sendReply`->`sendReplyInternal`->`sendErrorWrappedInternal`->`sendReplyInternal`->`sendErrorWrappedInternal` when even the too big exception is too big according to the limit (which could only happen for an insanely small limit).

The new version still handles that recursion. The only behavior change is that we don't enforce the limit on Thrift exceptions coming from thrift.

Reviewed By: yfeldblum

Differential Revision: D13944351

fbshipit-source-id: a6ed38aeeb60fb7b7b2275e9a2b1ceb90d1379e2