Re-organize TConnectionContext/Cpp2ConnContext/Cpp2RequestContext
Summary:
`TConnectionContext` is a base context class that we inherited from Thrift1. Both `Cpp2RequestContext` and `Cpp2ConnContext` inherit from it, however most of the fields are not relevant for `Cpp2RequestContext`, because it forwards for all connection related information to the ConnContext.

We still need `TConnectionContext`, because it's used as a common interface in all of our interfaces (EventHandlers in particular). However, we never use `TConnectionContext` on its own and always instantiate one of its subclasses.

This diff slightly reorganizes these 3 classes - all connection related information is moved to ConnContext. This allows to reduce `Cpp2RequestContext` in size by ~2x by eliminating unused fields. It also allows to have a separate thin class for client side context (next diff).

Reviewed By: jmswen

Differential Revision: D13960880

fbshipit-source-id: e2501f9ec6bddf938079cfd3c88d364c951adf9e