Support callbacks for onWriteSuccess
Summary:
Title. This allows users/libraries/frameworks to correctly measure when a request was successfully submitted to the kernel. Note that the API is somewhat asymmetric in that there is no `onWriteError()` API; this is because errors are currently handled by catching exceptions in `addTaskFinally`.

Note that I have only integrated the callback with `REQUEST_RESPONSE` and `FNF` APIs; I will do `REQUEST_STREAM` in a follow-up diff if the general approach in this diff seems fine.

Reviewed By: vitaut

Differential Revision: D13864026

fbshipit-source-id: d31f316f06ff9fb293bbfd943bb3603dd00784e3