allow py3.name to not be conditional
Summary:
Before we only allowed a py3.rename when the existing name was in a
fixed list of globally reserved words. like "async" or "from".

Instead of comming up with when a rename is allowed, allow it anywhere.

Also changing the annotation from py3.rename to py3.name to match the cpp.name

(Note: this ignores all push blocking failures!)

Reviewed By: yfeldblum, danshiovitz

Differential Revision: D14028543

fbshipit-source-id: 2819491f3395d03488f49267cd24d77170c95df0