Improve indirection handling, part 2
Summary:
My previous diff added support for indirection with `IOBuf` as underlying type, but it didn't properly handle the case of getter-like accessors that return a reference. In such scenario, we deduce `folly::IOBuf&` as a type and try to instantiate a wrong StringTraits template.

This diff fixes handles that corner case by adding `std::remove_reference`.

Reviewed By: yfeldblum

Differential Revision: D14018413

fbshipit-source-id: ba204745ba7debc9e53ab20c753d5b8c445fd727