Integrate rocket server with ThriftServer
Summary:
This diff integrates `RocketServerConnection` into Thrift server, providing a more performant alternative to `ManagedRSocketConnection`. Note that I still need to address how to roll this out safely (the options seem to be to guard the usage of the new implementation via a gflag or to use a different transport protocol string altogether).

Note that on top of this diff, I am still work on making all the header protocol functionality of `Cpp2Connection`/`Cpp2Request` (e.g., `isOverloaded()`, respecting various server options, ...) available in `RocketServerConnection`.

Reviewed By: spalamarchuk

Differential Revision: D13104078

fbshipit-source-id: c646c74381c049f527fb8bfec134577e32d0d84f