Bugfix: destroy context unconditionally on stream cancellation
Summary: This diff relates to a race between cancellation messages coming in from the client on the IO thread, and `onNext`/`onError`/`onComplete` callbacks, which are scheduled asynchronously on the IO thread from a ThreadManager thread (before the client cancellation frames are processed). As soon as `cancel()` is processed, the associated `RocketServerFrameContext` should immediately be destroyed/unusable. For example, the `cancel()` may occur when the connection is closed. This diff 1) destroys the context unconditionally on cancellation, and 2) safeguards callbacks from potentially dereferencing a null context.

Reviewed By: yfeldblum

Differential Revision: D14037230

fbshipit-source-id: 8f749cfa22e6055c8926c80fa26ed3c88f46664f