Provide BadEnum with __hash__ function that is consistent with values, not memory location
Summary:
Thrift Structs that contain a BadEnum, had a hash that was based on the memory location of the BadEnum instead of its
values.  This bug is as old as enum support.

Moving to a defined `__hash__` function allows us to base the hash on the immutable values of the BadEnum instance.  Instead of the default behavior of python `object`

Reviewed By: danshiovitz

Differential Revision: D14138631

fbshipit-source-id: c89c86bc5673e61ae650dd706841ea56943e26c5