AdmissionStrategy, aggregate stats for PriorityAdmissionStrategy
Summary:
Currently the PriorityAdmissionStrategy is reporting metrics on a per client_id + priority number basis. It means that if one client_id has a priority of 256, it will generate 256*5 metrics, which is unreadable!

This diff aggregates the metrics on a per client_id basis. Summing the metric per name.

Note we also sum the metric `*max` and `*limit` as the actual maximum queue is the sum of all the small queues.

Reviewed By: yfeldblum, vitaut

Differential Revision: D13310856

fbshipit-source-id: ccdf8f767cd7dd741b6722db6208fc80c1243efd