Thrift py3 client unittest oss (#312)
Summary:
* Create thriftcpp2_shared.so

Number of the symbols used from within thrift and folly need to be
singletons. In the initial prototype, had been statically linking
everything, which works until two thrift modules need to be loaded
simultaneously, a significant example of this is running unit tests.

All code for the shared library needs to be compiled with position
independent code -fPIC, so cpp and cpp2 static libs are replicated
with _pic variants.

During this split got run-time link failures on

apache::thrift::JSONProtocolReaderCommon::kEscapeChars

This is because we cannot extern a constexpr, as it would need to be
evaluated at compile-time. To work around this have switched it from
constexpr to const.

* When building using CMAKE_PREFIX_PATH needs to be passed to setup.py

* Automatically create __init__.py when generating thrift-py3 library

* Big shuffle of all the libraries include order, as starting to actually
be used at runtime, resulting in unresolved symbols.

* Add build_ext into the install function, to ensure correct config passed when
  'setuptools install' decides to re-evaluate the compile
Pull Request resolved: https://github.com/facebook/fbthrift/pull/312

Reviewed By: yfeldblum

Differential Revision: D13987733

Pulled By: calebmarchent

fbshipit-source-id: de988c1cf00c8446f59ffdc2efa8856a943eaeda